+++ src/core/composeRefs/composeRefs.js
@@ -1,0 +1,9 @@
+import setRefs from '../setRefs'
+
+function composeRefs(handlers = []) {
+  return element => {
+    setRefs(element, handlers)
+  }
+}
+
+export default composeRefs