+++ src/core/setRefs/index.d.ts
@@ -1,1 +1,2 @@
 export { default } from './setRefs'
+export * from './setRefs'